Update so this works out of box with 28.0.50
The newest version seems to have broken the init.el part around the
package-initialize. I removed the logic there and all seems to now work as
intended. I did not test earlier versions of emacs against this change. Sorry.

Seems the easiest way to fix things if the init.el errors-out for some reason
is to just do a M-x list-packages, install org and use-package, then re-enter
emacs. This has seemed to get me over the hump on occasion