Back to white background - moved customs to config
Lots of custom settings were separated from their use-package instantiation.
To make it easier to change the configuration, the variables that aren't
really meant to be changed via 'customize' have been moved into the
configuration org-mode file. Some settings are still left in the custom.el
file but many have been moved so they are co-located with the package usage.